# specs/1-jwt-auth-security/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo Application Authentication API
  version: 1.0.0
  description: JWT-based authentication and protected task management API

servers:
  - url: https://api.todo-app.com/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Development server

security:
  - bearerAuth: []

paths:
  /auth/signup:
    post:
      summary: Create new user account
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for authenticated requests
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid request data
        '409':
          description: Email already in use

  /auth/signin:
    post:
      summary: Authenticate user and get JWT token
      operationId: signin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for authenticated requests
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
        '404':
          description: User not found

  /tasks:
    get:
      summary: List user's tasks
      operationId: listTasks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized (missing or invalid token)
    post:
      summary: Create new task for authenticated user
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '400':
          description: Invalid task data

  /tasks/{task_id}:
    get:
      summary: Get specific task (ownership enforced)
      operationId: getTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details
          content:
            application/json:
              $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (task not owned by user)
        '404':
          description: Task not found
    put:
      summary: Update task (ownership enforced)
      operationId: updateTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (task not owned by user)
        '404':
          description: Task not found
    delete:
      summary: Delete task (ownership enforced)
      operationId: deleteTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (task not owned by user)
        '404':
          description: Task not found

  /tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion status (ownership enforced)
      operationId: toggleTaskCompletion
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (task not owned by user)
        '404':
          description: Task not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time

    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    TaskCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        completed:
          type: boolean
          default: false
      required:
        - title

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean